\name{Yraw}
\alias{Yraw}
\title{Example matrix of physical activity features}

\description{
\code{Yraw} is an example numeric matrix of physical activity features to be 
harmonized. Each row corresponds to one observation for a subject on a given day, 
and each column represents a distinct physical activity feature.
}
